enhancement: initial revamp on proposal content component
