+++ packages/react-app-revamp/components/_pages/Submission/Desktop/index.tsx
@@ -1,5 +1,5 @@
 import DialogModalProposal from "@components/_pages/DialogModalProposal";
-import { Proposal } from "@components/_pages/ProposalContent";
+import { Proposal } from "@components/_pages/";
 import { FC } from "react";
 
 interface SubmissionPageDesktopLayoutProps {
